Se entrega la página hotel-detail incluyendo el mapa de google conectado con la api de google maps
