+++ front/.env
@@ -1,1 +1,1 @@
-NEXT_PUBLIC_GOOGLE_MAPS_API_KEY = 'AIzaSyAlyhB1KOER9nmeyRt7XTZyVvVMz3ikb1c'
+NEXT_PUBLIC_GOOGLE_MAPS_API_KEY = ''
