+++ front/.env.example
@@ -1,0 +1,1 @@
+NEXT_PUBLIC_GOOGLE_MAPS_API_KEY = 'abcdeferg'
