+++ front/package.json
@@ -10,6 +10,7 @@
   },
   "dependencies": {
     "@googlemaps/js-api-loader": "^1.16.6",
+    "axios": "^1.7.2",
     "dotenv": "^16.4.5",
     "formik": "^2.4.6",
     "next": "14.2.5",
